[fix] 修复亿联银行对于 SDK 安全性检测的 bug

See merge request sensors-analytics/sdk/sa-sdk-ios!510

(cherry picked from commit 7048f49e829fbed463cb3cbeedddb234420271c1)

8e12e709 [fix] 修复亿联银行对于 SDK 安全性检测的 bug
5b4ec495 恢复现有逻辑
ee138934 修改返回类型不一致的问题
7a17d011 直接强转返回值类型

+++ SensorsAnalyticsSDK/SensorsAnalyticsSDK.m
@@ -721,7 +721,7 @@ static SensorsAnalyticsSDK *sharedInstance = nil;
 
 - (UInt64)getMaxCacheSize {
     @synchronized(self) {
-        return self.configOptions.maxCacheSize;
+        return (UInt64)self.configOptions.maxCacheSize;
     };
 }
